+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## 1.5.3 - Upcoming
+
+* Preserve explicit `null` JSON response properties
+
 ## 1.5.2 - 2026-05-22
 
 * Fix request serialization for validation-only `additionalParameters`